requirements.txt
pillow
streamlit
pillow
google-genai
